+++ cmd/influx/setup.go
@@ -7,6 +7,7 @@ import (
 	"strconv"
 	"strings"
 
+	"github.com/influxdata/influxdb"
 	platform "github.com/influxdata/influxdb"
 	"github.com/influxdata/influxdb/cmd/influx/internal"
 	"github.com/influxdata/influxdb/http"
@@ -236,28 +237,42 @@ You have entered:
 	}
 }
 
-var (
-	errPasswordIsNotMatch = fmt.Errorf("passwords do not match")
-	errPasswordIsTooShort = fmt.Errorf("passwords is too short")
-)
+func errSecretIsNotMatch(title string) error {
+	return fmt.Errorf(title + "s do not match")
+}
+
+func errSecretIsTooShort(title string) error {
+	return &influxdb.Error{
+		Code: influxdb.EUnprocessableEntity,
+		Msg:  title + " is too short",
+	}
+}
+
+func getSecret(ui *input.UI) (secret string) {
+	return getSecretInput(ui, false, "secret")
+}
 
 func getPassword(ui *input.UI, showNew bool) (password string) {
+	return getSecretInput(ui, showNew, "password")
+}
+
+func getSecretInput(ui *input.UI, showNew bool, title string) (secret string) {
 	newStr := ""
 	if showNew {
 		newStr = " new"
 	}
 	var err error
-enterPasswd:
-	query := string(promptWithColor("Please type your"+newStr+" password", colorCyan))
+enterSecret:
+	query := string(promptWithColor("Please type your"+newStr+" "+title, colorCyan))
 	for {
-		password, err = ui.Ask(query, &input.Options{
+		secret, err = ui.Ask(query, &input.Options{
 			Required:  true,
 			HideOrder: true,
 			Hide:      true,
 			Mask:      false,
 			ValidateFunc: func(s string) error {
 				if len(s) < 8 {
-					return errPasswordIsTooShort
+					return errSecretIsTooShort(title)
 				}
 				return nil
 			},
@@ -265,25 +280,25 @@ enterPasswd:
 		switch err {
 		case input.ErrInterrupted:
 			os.Exit(1)
-		case errPasswordIsTooShort:
-			ui.Writer.Write(promptWithColor("Password too short - minimum length is 8 characters!", colorRed))
-			goto enterPasswd
 		default:
-			if password = strings.TrimSpace(password); password == "" {
+			if influxdb.ErrorCode(err) == influxdb.EUnprocessableEntity {
+				ui.Writer.Write(promptWithColor(strings.ToTitle(title)+" too short - minimum length is 8 characters!\n\r", colorRed))
+				goto enterSecret
+			} else if secret = strings.TrimSpace(secret); secret == "" {
 				continue
 			}
 		}
 		break
 	}
-	query = string(promptWithColor("Please type your"+newStr+" password again", colorCyan))
+	query = string(promptWithColor("Please type your"+newStr+" "+title+" again", colorCyan))
 	for {
 		_, err = ui.Ask(query, &input.Options{
 			Required:  true,
 			HideOrder: true,
 			Hide:      true,
 			ValidateFunc: func(s string) error {
-				if s != password {
-					return errPasswordIsNotMatch
+				if s != secret {
+					return errSecretIsNotMatch(title)
 				}
 				return nil
 			},
@@ -294,12 +309,12 @@ enterPasswd:
 		case nil:
 			// Nothing.
 		default:
-			ui.Writer.Write(promptWithColor("Passwords do not match!\n", colorRed))
-			goto enterPasswd
+			ui.Writer.Write(promptWithColor(strings.ToTitle(title)+"s do not match!\n", colorRed))
+			goto enterSecret
 		}
 		break
 	}
-	return password
+	return secret
 }
 
 func getInput(ui *input.UI, prompt, defaultValue string) string {